// src/clients/web/src/types/protocol_buf_generated/boxteam/workspace/v2/file_events_pb.d.ts
// @generated by protoc-gen-es v2.14.0 with parameter "target=js+dts,import_extension=.js"
// @generated from file boxteam/workspace/v2/file_events.proto (package boxteam.workspace.v2, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file boxteam/workspace/v2/file_events.proto.
 */
export declare const file_boxteam_workspace_v2_file_events: GenFile;

/**
 * @generated from message boxteam.workspace.v2.WorkspaceFileChange
 */
export declare type WorkspaceFileChange = Message<"boxteam.workspace.v2.WorkspaceFileChange"> & {
  /**
   * @generated from field: string kind = 1;
   */
  kind: string;

  /**
   * @generated from field: string path = 2;
   */
  path: string;
};

/**
 * Describes the message boxteam.workspace.v2.WorkspaceFileChange.
 * Use `create(WorkspaceFileChangeSchema)` to create a new message.
 */
export declare const WorkspaceFileChangeSchema: GenMessage<WorkspaceFileChange>;

/**
 * @generated from message boxteam.workspace.v2.WorkspaceFileChangeBatch
 */
export declare type WorkspaceFileChangeBatch = Message<"boxteam.workspace.v2.WorkspaceFileChangeBatch"> & {
  /**
   * @generated from field: repeated boxteam.workspace.v2.WorkspaceFileChange changes = 1;
   */
  changes: WorkspaceFileChange[];

  /**
   * @generated from field: bool overflow = 2;
   */
  overflow: boolean;
};

/**
 * Describes the message boxteam.workspace.v2.WorkspaceFileChangeBatch.
 * Use `create(WorkspaceFileChangeBatchSchema)` to create a new message.
 */
export declare const WorkspaceFileChangeBatchSchema: GenMessage<WorkspaceFileChangeBatch>;

/**
 * @generated from message boxteam.workspace.v2.SseError
 */
export declare type SseError = Message<"boxteam.workspace.v2.SseError"> & {
  /**
   * @generated from field: string message = 1;
   */
  message: string;
};

/**
 * Describes the message boxteam.workspace.v2.SseError.
 * Use `create(SseErrorSchema)` to create a new message.
 */
export declare const SseErrorSchema: GenMessage<SseError>;